requirements.txt
paho-mqtt==1.6.1
xmltodict==0.13.0
